Добавил проверку на пустые поля у форм
