+++ packages/client/src/components/forum-components/comments-reply-modal/types.ts
@@ -1,0 +1,9 @@
+import { ChangeEvent, FormEvent } from 'react'
+
+export type TCommentsReplyModal = {
+  isOpenModal: boolean
+  handleCloseModal: () => void
+  handleSendReply: (e: FormEvent<HTMLFormElement>) => void
+  handleCancel: () => void
+  handleChangeMessage: (e: ChangeEvent<HTMLInputElement>) => void
+}
